Add utilities file and refactor lib directory
