+++ lib/family.js
@@ -1,42 +1,17 @@
-import fs from 'fs';
 import path from 'path';
-import matter from 'gray-matter';
-import remark from 'remark';
-import html from 'remark-html';
+import { getSortedData, getAllIds, getData } from '../utilities';
 
-//set filename path with process.cwd()
-const familyDirectory = path.join(process.cwd(), 'family')
+const familyDirectory = path.join(process.cwd(), 'family');
 
 export function getSortedFamilyData() {
-  const fileNames = fs.readdirSync(familyDirectory)
-  return fileNames.map(fileName => {
-    // Remove ".png" from file name to get id
-    const id = fileName.replace(/\.md$/, '')
+  return getSortedData(familyDirectory);
+};
 
-    // Read markdown file as string
-    const fullPath = path.join(familyDirectory, fileName)
-    const fileContents = fs.readFileSync(fullPath, 'utf8')
+export function getAllFamilyIds() {
+  return getAllIds(familyDirectory);     
+};
 
-    // Use gray-matter to parse the post metadata section
-    const matterResult = matter(fileContents)
-
-    // Combine the data with the id
-    return {
-      id,
-      ...matterResult.data
-    }
-  })
-}
-
-// export function getAllFamilyIds() {
-//   const fileNames = fs.readdirSync(familyDirectory);
-
-//   return fileNames.map(fileName => {
-//     return {
-//       params: {
-//         id: fileName.replace(/\.png$/, '')
-//       }
-//     }
-//   })
-// }
+export async function getFamilyData(id) {
+  return getData(familyDirectory, id)
+};
